Fix swiping through elements be removing elements update
